rename: style 관련 도메인 member 패키지로 이동
