+++ src/main/java/com/mfc/memberservice/member/infrastructure/FavoriteStyleRepository.java
@@ -1,10 +1,10 @@
-package com.mfc.memberservice.style.infrastructure;
+package com.mfc.memberservice.member.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-import com.mfc.memberservice.style.domain.FavoriteStyle;
+import com.mfc.memberservice.member.domain.FavoriteStyle;
 
 public interface FavoriteStyleRepository extends JpaRepository<FavoriteStyle, Long> {
 	@Modifying(clearAutomatically = true)
